+++ src/main/java/com/multi/shop/api/multi_shop_api/products/entities/dtos/UpdateProductDTO.java
@@ -1,6 +1,5 @@
 package com.multi.shop.api.multi_shop_api.products.entities.dtos;
 
-import java.math.BigDecimal;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -24,7 +23,7 @@ public class UpdateProductDTO {
     private String description;
     
     @NotNull(message = "{NotBlank.validation.text}")
-    private BigDecimal price;
+    private Long price;
 
     @JsonIgnoreProperties("id")
     private List<Image> productImages;
@@ -42,7 +41,7 @@ public class UpdateProductDTO {
         productImages = new ArrayList<> ();
     }
 
-    public UpdateProductDTO(String productName, String description, BigDecimal price, List<ProductCategory> categories) {
+    public UpdateProductDTO(String productName, String description, Long price, List<ProductCategory> categories) {
         this.productName = productName;
         this.description = description;
         this.price = price;
@@ -65,11 +64,11 @@ public class UpdateProductDTO {
         this.description = description;
     }
 
-    public BigDecimal getPrice() {
+    public Long getPrice() {
         return price;
     }
 
-    public void setPrice(BigDecimal price) {
+    public void setPrice(Long price) {
         this.price = price;
     }
